Move all models in unit tests to separate files for easier navigation
