+++ src/Catel.Test/Catel.Test.Shared/Catel.Test.Shared.projitems
@@ -32,9 +32,35 @@
     <Compile Include="$(MSBuildThisFileDirectory)Data\BoxingCacheFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\CatelTypeInfoFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ChangeNotificationWrapperFacts.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ClassWithoutPropertiesA.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ClassWithoutPropertiesB.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ClassWithValidator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\Model.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ModelA.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ModelB.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ModelC.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithValidation.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\Child.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ChildAwareModelBaseFacts.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ClassWithPropertiesRegisteredByNonMagicStringOverload.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\CompositeValidatorFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\CompositeValidatorProviderFacts.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\NonSerializableClass.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithNonSerializableMembers.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithoutDefaultValues.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ComputerSettingsWithXmlMappings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithPrivateConstructor.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\Parent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\Gender.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithCustomType.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithPrivateMembers.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithXmlMappings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ComputerSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\IniEntryType.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\IniFile.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ExtendedIniEntry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\IniEntry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\DynamicObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\EventArgs\AdvancedPropertyChangedEventArgsFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Exceptions\InvalidPropertyExceptionFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\Exceptions\InvalidPropertyValueExceptionFacts.cs" />
@@ -48,9 +74,12 @@
     <Compile Include="$(MSBuildThisFileDirectory)Data\ModelBaseFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ModelBaseFacts.editableobject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ModelBaseFacts.serialization.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ObjectWithDefaultValues.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\TestValidatorProvider.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\TestValidator.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ValidatableModelBaseFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ModelBaseTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Data\ModelBaseTestHelper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\Helpers\ModelBaseTestHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\PropertyBagFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\PropertyDataFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\PropertyDataManagerFacts.cs" />
@@ -66,6 +95,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Data\ValidationContextFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ValidationResultFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Data\ValidationSummaryFacts.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Data\TestClasses\ValidationTestModel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DisposableTokenFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)EnumFacts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExceptionHandling\ExceptionExtensionsTests.cs" />
