+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/ClassWithPropertiesRegisteredByNonMagicStringOverload.cs
@@ -1,0 +1,46 @@
+namespace Catel.Test.Data
+{
+    using Catel.Data;
+
+    public class ClassWithPropertiesRegisteredByNonMagicStringOverload : ModelBase
+    {
+        public static readonly PropertyData StringPropertyProperty = RegisterProperty<ClassWithPropertiesRegisteredByNonMagicStringOverload, string>(instance => instance.StringProperty);
+        
+        public static readonly PropertyData StringPropertyWithSpecifiedDefaultValueProperty = RegisterProperty<ClassWithPropertiesRegisteredByNonMagicStringOverload, string>(instance => instance.StringPropertyWithSpecifiedDefaultValue, "NonNullOrEmptyDefaultValue");
+
+        public static readonly PropertyData IntPropertyWithPropertyChangeNoticationProperty = RegisterProperty<ClassWithPropertiesRegisteredByNonMagicStringOverload, int>(instance => instance.IntPropertyWithPropertyChangeNotication , default(int), (s, e) => s.OnIntPropertyWithPropertyChangeNoticationChanged());
+
+        public static readonly PropertyData IntPropertyExcludedFromSerializationAndBackupProperty = RegisterProperty<ClassWithPropertiesRegisteredByNonMagicStringOverload, int>(instance => instance.IntPropertyExcludedFromSerializationAndBackup, default(int), null, false, false);
+
+        public int IntPropertyExcludedFromSerializationAndBackup
+        {
+            get { return GetValue<int>(IntPropertyExcludedFromSerializationAndBackupProperty); }
+            set { SetValue(IntPropertyExcludedFromSerializationAndBackupProperty, value); }
+        }
+
+        public string StringPropertyWithSpecifiedDefaultValue
+        {
+            get { return GetValue<string>(StringPropertyWithSpecifiedDefaultValueProperty); }
+            set { SetValue(StringPropertyWithSpecifiedDefaultValueProperty, value); }
+        }
+
+        public string StringProperty
+        {
+            get { return GetValue<string>(StringPropertyProperty); }
+            set { SetValue(StringPropertyProperty, value); }
+        }
+
+        public int IntPropertyWithPropertyChangeNotication
+        {
+            get { return GetValue<int>(IntPropertyWithPropertyChangeNoticationProperty); }
+            set { SetValue(IntPropertyWithPropertyChangeNoticationProperty, value); }
+        }
+
+        public int IntPropertyWithPropertyChangeNoticationsCount { get; private set; }
+
+        private void OnIntPropertyWithPropertyChangeNoticationChanged()
+        {
+            this.IntPropertyWithPropertyChangeNoticationsCount++;
+        }
+    }
+}
