+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/Gender.cs
@@ -1,0 +1,9 @@
+﻿namespace Catel.Test.Data
+{
+    public enum Gender
+    {
+        Male,
+
+        Female
+    }
+}
