+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/IniEntryType.cs
@@ -1,0 +1,12 @@
+namespace Catel.Test.Data
+{
+    /// <summary>
+    /// Just a dummy enum to test enum (de)serialization.
+    /// </summary>
+    public enum IniEntryType
+    {
+        Public,
+
+        Private
+    }
+}
