+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/NonSerializableClass.cs
@@ -1,0 +1,6 @@
+namespace Catel.Test.Data
+{
+    public class NonSerializableClass
+    {   
+    }
+}
