+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/ObjectWithNonSerializableMembers.cs
@@ -1,0 +1,21 @@
+namespace Catel.Test.Data
+{
+    using Catel.Data;
+
+    public class ObjectWithNonSerializableMembers : ModelBase
+    {
+        /// <summary>
+        ///   Gets or sets a non-serializable value.
+        /// </summary>
+        public NonSerializableClass NonSerializableValue
+        {
+            get { return GetValue<NonSerializableClass>(NonSerializableValueProperty); }
+            set { SetValue(NonSerializableValueProperty, value); }
+        }
+
+        /// <summary>
+        ///   Register the NonSerializableValue property so it is known in the class.
+        /// </summary>
+        public static readonly PropertyData NonSerializableValueProperty = RegisterProperty("NonSerializableValue", typeof(NonSerializableClass), null);
+    }
+}
