+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/TestValidatorProvider.cs
@@ -1,0 +1,15 @@
+﻿namespace Catel.Test.Data
+{
+    using System;
+    using Catel.Data;
+
+    public class TestValidatorProvider : ValidatorProviderBase
+    {
+        protected override IValidator GetValidator(Type targetType)
+        {
+            Argument.IsNotNull("targetType", targetType);
+
+            return new TestValidator();
+        }
+    }
+}
